Split model test utils into common and mapping-specific
